requirements.txt
# Web Framework
fastapi==0.68.1
uvicorn==0.15.0
jinja2==3.0.1
python-multipart==0.0.5
aiofiles==0.7.0

# Data Processing
pyyaml==5.4.1
feedparser==6.0.10
beautifulsoup4==4.9.3
aiohttp==3.8.1
pydantic==1.8.2

# Development Tools
python-dotenv==0.19.0
virtualenv==20.13.0

# Logging & Monitoring
loguru==0.5.3
prometheus-client==0.12.0

# Testing
pytest==6.2.5
pytest-asyncio==0.16.0

# Code Quality
black==21.12b0
isort==5.10.1
flake8==4.0.1

# System
psutil==5.8.0
typing-extensions==4.0.1 